website/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
psycopg2-binary==2.9.9
werkzeug==2.3.7